Python: Add API function to reorder curves
The order of curves in a Grease Pencil drawing defines the
order in wich they are rendered, so it's important to be
able to change the order. Currently, there is not easy way
to do this from python.

This adds a new `reorder_curves` and `reorder_strokes`
function to reorder the curves in the curves geometry.

The function takes in an array of indices that are the
new indices for the curves (e.g. an index map).

`reorder_curves(new_indices=[...])`

Pull Request: https://projects.blender.org/blender/blender/pulls/130331