Geometry Nodes: support text overlay when viewing matrix attributes
This implements a basic text overlay for matrix attributes. The matrix is split into location,
rotation and scale to make the values more useful. There are probably ways to present
this in a more beautiful way, but for now this is already much more useful compared to
showing nothing. Also, previously, this would hit an assert in a debug build.

Co-authored-by: Hans Goudey <hans@blender.org>
Pull Request: https://projects.blender.org/blender/blender/pulls/131301