Fix #127511: inconsistent tangents in Set Position node
The issue was that the `handle_left`, `handle_right` and `position` attribute
are updated one after another and the input field is evaluated again for each
attribute. If the handle positions are updated first, then the tangent will
change in consecutive evaluations.

The fix is to set the three attributes at the same time and to evaluate the
inputs only once.

Pull Request: https://projects.blender.org/blender/blender/pulls/131609